Fix for AMS 0.19, remove No Override
The introduction of tipc broke Nro2Nsp. The updated versions are made with hacBrewPack, which compiles the stub screen as the main executable in ExeFS, reducing the size by about 600kb. The non-override version was removed because SX OS is dead, and it's broken on AMS because of the aforementioned issue.