Enable automatic video capture
Specify "system_resource_size": "0x1FE00000" in your npdm json when enabling auto recording or AM gets upset. 0x1FE00000 is the max according to SwitchBrew. This isn't required for manual recording, but setting it up is left to the running homebrew, and my guess is most homebrew doesn't bother.